+++ src/tests/correctness.rs
@@ -671,7 +671,7 @@ mod hashmap_test {
 #[cfg(test)]
 mod hashindex_test {
     use crate::ebr;
-    use crate::hash_index::Visitor;
+    use crate::hash_index::{ModifyAction, Visitor};
     use crate::HashIndex;
     use proptest::strategy::{Strategy, ValueTree};
     use proptest::test_runner::TestRunner;
@@ -941,6 +941,85 @@ mod hashindex_test {
                 removed.store(i, Release);
             }
             thread_handle.join().unwrap();
+        }
+    }
+
+    #[cfg_attr(miri, ignore)]
+    #[tokio::test(flavor = "multi_thread", worker_threads = 8)]
+    async fn modify() {
+        let hashindex: Arc<HashIndex<usize, usize>> = Arc::new(HashIndex::default());
+        for _ in 0..256 {
+            let num_tasks = 8;
+            let workload_size = 256;
+            let mut task_handles = Vec::with_capacity(num_tasks);
+            let barrier = Arc::new(AsyncBarrier::new(num_tasks));
+            for task_id in 0..num_tasks {
+                let barrier_clone = barrier.clone();
+                let hashindex_clone = hashindex.clone();
+                task_handles.push(tokio::task::spawn(async move {
+                    barrier_clone.wait().await;
+                    let range = (task_id * workload_size)..((task_id + 1) * workload_size);
+                    for id in range.clone() {
+                        let result = hashindex_clone.insert_async(id, id).await;
+                        assert!(result.is_ok());
+                        if id % 4 == 0 {
+                            assert!(!hashindex_clone.modify(&id, |_, v| if *v == id {
+                                ModifyAction::Keep
+                            } else {
+                                ModifyAction::Update(0)
+                            }));
+                        } else {
+                            assert!(
+                                !hashindex_clone
+                                    .modify_async(&id, |_, v| if *v == id {
+                                        ModifyAction::Keep
+                                    } else {
+                                        ModifyAction::Update(0)
+                                    })
+                                    .await
+                            );
+                        }
+                    }
+                    for id in range.clone() {
+                        hashindex_clone.read(&id, |k, v| assert_eq!(k, v));
+                        if id % 4 == 0 {
+                            assert!(hashindex_clone.modify(&id, |_, v| if *v == id {
+                                Some(Some(usize::MAX))
+                            } else {
+                                None
+                            }));
+                        } else {
+                            assert!(
+                                hashindex_clone
+                                    .modify_async(&id, |_, v| if *v == id {
+                                        Some(Some(usize::MAX))
+                                    } else {
+                                        None
+                                    })
+                                    .await
+                            );
+                        }
+                    }
+                    for id in range.clone() {
+                        hashindex_clone.read(&id, |_, v| assert_eq!(*v, usize::MAX));
+                        if id % 4 == 0 {
+                            assert!(hashindex_clone.modify(&id, |_, _| Some(None)));
+                        } else {
+                            assert!(
+                                hashindex_clone
+                                    .modify_async(&id, |_, _| ModifyAction::Remove)
+                                    .await
+                            );
+                        }
+                    }
+                }));
+            }
+
+            for r in futures::future::join_all(task_handles).await {
+                assert!(r.is_ok());
+            }
+
+            assert_eq!(hashindex.len(), 0);
         }
     }